Fixed a couple things with the migration:
When migrating an actor, we were setting a constant of actorData to the
data of the actor.  Then we were migrating the actors items and then
using actorData to update the actor.  The problem is, actorData HAS all
of the old item data in it!

Moved constant to after the item update.

Fixed the type checking on when it should migrate character data vs all
actors.  It was incorrectly referencing actor.type instead of
actorData.type.

Added code to clean the actors of the old unused data points after
advice from Moo Man on how to actually remove data points from the
actors.