Did some re-structuring of the flow to get DSN to work.  When someone issues a /red command:
- The chatMessage hook gets called, /red is stripped from the message and the rest is passed to CPRChat.HandleCPRCommand()
- This looks at the data and tries to match it first to a XdY formula, and if it can not, it uses 1d10, otherwise it uses the passed formula (more on this in a bit)
- It looks to see if modifiers were passed as +/- and appends those to the formula
- Finally if we have a formula, it calls CPRRoll to create the roll, rolls it and then renders the chat card

In order to make the /red command more universal and allow the passing of other formulas (such as 1d6), I updated the constructor to store the die type on the CPRoll object and then
updated the HBS to use this as the class for displaying the dice.  All we need now is the SVG's for all of the other dice and we can then be able to display ANY dice to chat.