This addresses the issue mentioned:
Opening actor sheets for some actors produce, "CPR WRN | Improper use of the filter helper. This should not occur. Always provide an object list and not an undefined value. The following arguments were passed: objList = undefined, key = data.data.type, value = cyberware" It is possible this comes from the migration error(s).

The problem was that during the migration every item is re-created and Cyberware items specifically, has references between each other in the optionalIds datapoint which were not updated with the new item IDs.  This commit fixes that issue.