Add test to ensure packs are sorted alphabetically by key
Packs are horrible to work on so let's sort them by keyname.

This will mean that when exporting changes from within foundry you have
to run the file through the below before committing:

  `cat ${db} | jq --sort-keys -c > foo.db && cp foo.db ${db}`

The advantage to this is that you don't get arbitrary changes in diffs
if Foundry decides to mix up lines/ordering which should in theory make
compendium changes easier to review.