Items can be dragged from an actor sheet to the right side window which creates a new item in the world. There's a two problems when this happens, one affects all worlds, the other affects only a world that was migrated.
If the item is a ranged weapon and has ammo in it, the new item should have that ammo removed from it, it does not and still has references to the original ammo item.
If the item has an upgrade applied to it, the new item should not be upgraded and have all references to the original upgrade removed from it.  This seems to work on a `New World` however on a world that was `Migrated`, it seems to be keeping the item upgrade references.

This fixes these two issues.