Drag and drop changes a bit in v10 which caused dragging items from characters to the Foundry System to create world items to fail becuase we over-ride onDragStart and the Foundry onDrop was expecting a different dataset for dataTransfer of items.
Adjusted code to work with v10 for this and then had to go back and change the way we handled our onDrop() to work with the new format.  This required a change to the Container Actors as well which then revealed a completely different issue with them (inability to configure purchase types).

Therefore this commit has fixes for all of those things.