Split & ESLint gulpfile
The gulpfile.tasks was becoming a bit unweildy for adding more
functionality to, as such it has been split up and refactored to use
individual more focussed files in `gulp/`.

Also made sure we now pass ESlint and ensure the gulpfile and `gulp/`
files are included in the linting job.