Add 0.83.0 system.json to master to fix broken release
Users can't update the system within Foundry as the `system.json` is
missing from the master branch which is where Foundry looks for the
manifest.

This file should never exist in dev, this is purely to fix the upgrade
path for older version. It should be probably be remove once we're no
longer supporting 0.82